+++ Libplanet.Explorer.Tests/Queries/StateQueryTest.Legacy.cs
@@ -109,10 +109,10 @@ public partial class StateQueryTest
         Assert.Equal("ABC", currencyDict["ticker"]);
         Assert.Equal("84ba810ca5ac342c122eb7ef455939a8a05d1d40", currencyDict["hash"]);
         Assert.Equal(1, Assert.IsAssignableFrom<int>(totalSupplyDict["sign"]));
-        Assert.Equal(100, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["majorUnit"]));
-        Assert.Equal(0, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["minorUnit"]));
-        Assert.Equal("100", totalSupplyDict["quantity"]);
-        Assert.Equal("100 ABC", totalSupplyDict["string"]);
+        Assert.Equal(101, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["majorUnit"]));
+        Assert.Equal(23, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["minorUnit"]));
+        Assert.Equal("101.23", totalSupplyDict["quantity"]);
+        Assert.Equal("101.23 ABC", totalSupplyDict["string"]);
 
         result = await ExecuteQueryAsync<StateQuery>(@"
         {
@@ -268,11 +268,11 @@ public partial class StateQueryTest
             Assert.IsAssignableFrom<IDictionary<string, object>>(totalSupplyDict["currency"]);
         Assert.Equal("ABC", currencyDict["ticker"]);
         Assert.Equal("84ba810ca5ac342c122eb7ef455939a8a05d1d40", currencyDict["hash"]);
-        Assert.Equal(100, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["majorUnit"]));
+        Assert.Equal(101, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["majorUnit"]));
         Assert.Equal(1, Assert.IsAssignableFrom<int>(totalSupplyDict["sign"]));
-        Assert.Equal(0, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["minorUnit"]));
-        Assert.Equal("100", totalSupplyDict["quantity"]);
-        Assert.Equal("100 ABC", totalSupplyDict["string"]);
+        Assert.Equal(23, Assert.IsAssignableFrom<BigInteger>(totalSupplyDict["minorUnit"]));
+        Assert.Equal("101.23", totalSupplyDict["quantity"]);
+        Assert.Equal("101.23 ABC", totalSupplyDict["string"]);
 
         result = await ExecuteQueryAsync<StateQuery>(@"
         {